refactor: 헤더 뒤로가기 버튼을 router.back으로 변경
