+++ src/app/[lng]/(main)/profile/praise/components/PraiseHeader.client.tsx
@@ -1,20 +1,21 @@
+'use client';
 import { useTranslation } from '@/app/i18n/client';
 import { IconButton } from '@/components/Button';
 import { Header } from '@/components/Header';
 import { Icon } from '@/components/Icon';
-import Link from 'next/link';
+import { useRouter } from 'next/navigation';
 
-export default function Step1Header() {
+export default function PraiseHeader() {
   const { t } = useTranslation('profile');
+  const router = useRouter();
+
   return (
     <Header>
       <Header.Left className="px-4">
-        <Link href="/profile/setting">
-          <IconButton size="large">
-            <Icon id="24-close" />
-          </IconButton>
-        </Link>
-        <p className="text-subtitle-1">{t('settings.editProfile')}</p>
+        <IconButton size="large" onClick={() => router.back()}>
+          <Icon id="24-arrow_back" />
+        </IconButton>
+        <p className="text-subtitle-1">{t('home.praiseCount')}</p>
       </Header.Left>
     </Header>
   );
