+++ src/app/[lng]/(main)/profile/setting/components/SettingHeader.client.tsx
@@ -3,19 +3,18 @@ import { useTranslation } from '@/app/i18n/client';
 import { IconButton } from '@/components/Button';
 import { Header } from '@/components/Header';
 import { Icon } from '@/components/Icon';
-import Link from 'next/link';
+import { useRouter } from 'next/navigation';
 
 export default function SettingHeader() {
   const { t } = useTranslation('profile');
+  const router = useRouter();
 
   return (
     <Header>
       <Header.Left className="px-4">
-        <Link href="/profile">
-          <IconButton size="large">
-            <Icon id="24-arrow_back" />
-          </IconButton>
-        </Link>
+        <IconButton size="large" onClick={() => router.back()}>
+          <Icon id="24-arrow_back" />
+        </IconButton>
         <p className="text-subtitle-1">{t('settings.settings')}</p>
       </Header.Left>
     </Header>
