+++ src/app/[lng]/(main)/profile/setting/page.tsx
@@ -1,6 +1,6 @@
 import LinkSection from './components/LinkSection.server';
 import ProfileSection from './components/ProfileSection.client';
-import SettingHeader from './components/SettingHeader';
+import SettingHeader from './components/SettingHeader.client';
 import { RejectedFallback } from '@/components/ErrorBoundary';
 import { PageAnimation } from '@/components/PageAnimation';
 import { QueryAsyncBoundary } from '@suspensive/react-query';
